// yara-app-frontend/src/app/models/product-item.model.ts
export class ProductItem {
    productCode: String;
    productName: String;
    additionalInfo: String;
    categoryId: String;
    description: String;
    featuredProduct: Boolean;
    price: Number;
    quantity: Number;
    imageUrls: Array<String>;
    mainImageUrl: String;
    ratings: Number;
    returnable: Boolean;
    reviews: Array<String>;
    saleDiscount: Number;
    tags: Array<String>;
    specifications: Array<String>;
    variants: Array<String>;

}